Improve rendering performance
This greatly increases throughput when an application is generating data
faster than we can render it. We achieve this by reading up to 512kb of
output in a loop and then rendering the result.

Using this strategy, we actually improve the rendering throughput of
Kitty / Alacritty by around 16% despite adding an emulation layer